refactor(create): move suspense to layout
