+++ service/app/src/app/create/layout.tsx
@@ -1,29 +1,19 @@
 "use client"
 
-import { useSearchParams } from "next/navigation"
 import type { ReactNode } from "react"
-import { useEffect } from "react"
+import { Suspense } from "react"
 
-import { CreateGameProvider, gameStoreInstance } from "./store"
+import CreateGamePageSkeleton from "./loading"
+import { CreateGameProvider } from "./store"
 
 export default function CreateGameLayout({
   children,
 }: {
   children: ReactNode
 }) {
-  const searchParams = useSearchParams()
-  const gameId = searchParams.get("gameId")
-
-  useEffect(() => {
-    if (gameId) {
-      const loadData = async () => {
-        await gameStoreInstance.getState().loadGameData(gameId)
-      }
-      loadData()
-    } else {
-      gameStoreInstance.getState().reset()
-    }
-  }, [gameId])
-
-  return <CreateGameProvider>{children}</CreateGameProvider>
+  return (
+    <Suspense fallback={<CreateGamePageSkeleton />}>
+      <CreateGameProvider>{children}</CreateGameProvider>
+    </Suspense>
+  )
 }
